+++ src/bitmap/color_plate_scanner.cpp
@@ -583,6 +583,7 @@ namespace Invader {
 
     void ColorPlateScanner::generate_mipmaps(GeneratedBitmapData &generated_bitmap, std::int16_t mipmaps, ScannedColorMipmapType mipmap_type, std::optional<float> mipmap_fade_factor, const std::optional<ColorPlateScannerSpriteParameters> &sprite_parameters, std::optional<float> sharpen, std::optional<float> blur) {
         auto mipmaps_unsigned = static_cast<std::uint32_t>(mipmaps);
+        float fade = mipmap_fade_factor.value_or(0.0F);
         for(auto &bitmap : generated_bitmap.bitmaps) {
             std::uint32_t mipmap_width = bitmap.width;
             std::uint32_t mipmap_height = bitmap.height;
@@ -758,26 +759,6 @@ namespace Invader {
                         }
 
                         #undef INTERPOLATE_CHANNEL
-
-                        // Do fade-to-gray
-                        if(mipmap_fade_factor.has_value()) {
-                            // Get the old alpha
-                            std::uint8_t old_alpha = pixel.alpha;
-
-                            // Alpha blend to gray
-                            std::uint8_t alpha_delta = static_cast<std::uint8_t>(0xFF / max_mipmap_count);
-                            ColorPlatePixel FADE_TO_GRAY = { 0x7F, 0x7F, 0x7F, alpha_delta };
-                            pixel.alpha *= (1.0F - mipmap_fade_factor.value());
-                            pixel = pixel.alpha_blend(FADE_TO_GRAY);
-
-                            // Fade alpha to white
-                            if(UINT8_MAX - old_alpha > alpha_delta) {
-                                pixel.alpha = old_alpha + alpha_delta;
-                            }
-                            else {
-                                pixel.alpha = UINT8_MAX;
-                            }
-                        }
                     }
                 }
 
@@ -785,6 +766,48 @@ namespace Invader {
                 mipmap_height /= 2;
                 mipmap_width /= 2;
                 last_mipmap_offset = this_mipmap_offset;
+            }
+
+            // Do fade-to-gray for each mipmap
+            if(mipmap_fade_factor.has_value()) {
+                std::size_t mipmap_count = bitmap.mipmaps.size();
+                std::size_t mipmap_count_plus_one = mipmap_count + 1; // although Guerilla only mentions mipmaps in the fade-to-gray stuff, it includes the first bitmap in the calculation
+                float overall_fade_factor = static_cast<float>(mipmap_count_plus_one) - static_cast<float>(fade) * mipmap_count_plus_one;
+
+                for(std::size_t m = 0; m < mipmap_count; m++) {
+                    auto &mipmap = bitmap.mipmaps[m];
+
+                    // Iterate through each pixel
+                    ColorPlatePixel *first = bitmap.pixels.data() + mipmap.first_pixel;
+                    auto *last = first + mipmap.pixel_count;
+
+                    while(first < last) {
+                        std::uint8_t alpha_delta;
+
+                        // If we're fading to gray instantly, do that so we don't divide by 0
+                        if(fade >= 1.0F) {
+                            alpha_delta = 0xFF;
+                        }
+                        else {
+                            // Basically, a higher mipmap fade factor scales faster
+                            float gray_multiplier = static_cast<float>(m + 1) / overall_fade_factor;
+
+                            // If we go over 1, go to 1
+                            if(gray_multiplier > 1.0F) {
+                                gray_multiplier = 1.0F;
+                            }
+
+                            // Round
+                            float gray_multiplied = std::floor(0xFF * gray_multiplier + 0.5F);
+                            alpha_delta = static_cast<std::uint32_t>(gray_multiplied);
+                        }
+
+                        ColorPlatePixel FADE_TO_GRAY = { 0x7F, 0x7F, 0x7F, static_cast<std::uint8_t>(alpha_delta) };
+                        *first = first->alpha_blend(FADE_TO_GRAY);
+
+                        first++;
+                    }
+                }
             }
         }
     }