+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is used for recording Invader's changes. This changelog is based on [Keep a
 ### Fixed
 - invader-build: Fixed not fixing the render bounding radius if it was less than the bounding radius but non-zero.
 - invader-build: Fixed not setting the weight value for color change permutations in objects.
+- invader-bitmap: Fixed detail fade factor so it matches tool.exe's detail fade factor more closely if not exactly.
 
 ### Removed
 - invader-bitmap: Removed --filter-blur and --filter-sharpen. Tags that have these values set will still have the